refactor(tick): isolate serde format helpers

Copilot-Session: f9f5de95-4e26-4d37-ae68-ac5e22182bcb
