+++ crates/tick/src/fmt/ecmascript.rs
@@ -191,7 +191,7 @@ impl<'de> serde_core::Deserialize<'de> for EcmaScript {
     where
         D: serde_core::Deserializer<'de>,
     {
-        super::deserialize_from_str(deserializer)
+        super::serde::deserialize_from_str(deserializer)
     }
 }
